# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "News aggregator API",
        "title": "Prospero",
        "contact": {
            "name": "Vitalii Popov",
            "url": "https://www.linkedin.com/in/mskkote/",
            "email": "msk.vitaly@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/grandFilter": {
            "post": {
                "description": "Perform a grand filter search based on provided parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Perform grand filter search",
                "parameters": [
                    {
                        "description": "Array of search strings with \u0026\u0026 as the joining operator",
                        "name": "filterStrings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SearchString"
                            }
                        }
                    },
                    {
                        "description": "Array of search people",
                        "name": "filterPeople",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SearchPeople"
                            }
                        }
                    },
                    {
                        "description": "Array of search publishers",
                        "name": "filterPublishers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SearchPublishers"
                            }
                        }
                    },
                    {
                        "description": "Array of search countries",
                        "name": "filterCountry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SearchCountry"
                            }
                        }
                    },
                    {
                        "description": "Array of search categories",
                        "name": "filterCategories",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SearchCategory"
                            }
                        }
                    },
                    {
                        "description": "Array of search languages",
                        "name": "filterLanguages",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SearchLanguage"
                            }
                        }
                    },
                    {
                        "description": "Time filter",
                        "name": "filterTime",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SearchTime"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/searchPublisherWithHints/{search}": {
            "post": {
                "description": "Search publishers with hints based on provided search string",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search publishers with hints",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search string",
                        "name": "search",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.SearchCategory": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.SearchCountry": {
            "type": "object",
            "properties": {
                "country": {
                    "description": "название страны",
                    "type": "string"
                }
            }
        },
        "dto.SearchLanguage": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.SearchPeople": {
            "type": "object",
            "properties": {
                "fullName": {
                    "description": "имя",
                    "type": "string"
                }
            }
        },
        "dto.SearchPublishers": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "название издания",
                    "type": "string"
                }
            }
        },
        "dto.SearchString": {
            "type": "object",
            "properties": {
                "isExact": {
                    "type": "boolean"
                },
                "search": {
                    "type": "string"
                }
            }
        },
        "dto.SearchTime": {
            "type": "object",
            "properties": {
                "end": {
                    "description": "Окончание временного диапазона",
                    "type": "string"
                },
                "start": {
                    "description": "Начало временного диапазона",
                    "type": "string"
                }
            }
        },
...
    }
}